remove DottyIDE
