+++ language-server/src/dotty/tools/languageserver/DottyLanguageServer.scala
@@ -54,7 +54,6 @@ class DottyLanguageServer extends LanguageServer
   import lsp4j.jsonrpc.messages.{Either => JEither}
   import lsp4j._
 
-
   private[this] var rootUri: String = _
 
   private[this] var myClient: DottyClient = _
@@ -646,7 +645,7 @@ class DottyLanguageServer extends LanguageServer
 }
 
 object DottyLanguageServer {
-  /** Configuration file normally generated by the DottyIDEPlugin */
+
   final val IDE_CONFIG_FILE = ".dotty-ide.json"
 
   final val RENAME_OVERRIDDEN_QUESTION = "Do you want to rename the base member, or only this member?"
