--- vscode-dotty/.vscodeignore
@@ -1,11 +1,0 @@
-## This is used to exclude files from "vsce publish"
-target/**
-.vscode/**
-.vscode-test/**
-out/test/**
-out/.keep
-test/**
-src/**
-**/*.map
-.gitignore
-tsconfig.json
