--- vscode-dotty/package.json
@@ -1,138 +1,0 @@
-{
-  "name": "dotty",
-  "displayName": "Dotty Language Server",
-  "description": "IDE integration for Dotty, the experimental Scala compiler",
-  "version": "0.1.17-snapshot",
-  "license": "Apache-2.0",
-  "publisher": "lampepfl",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/lampepfl/dotty.git"
-  },
-  "icon": "images/dotty-logo.png",
-  "engines": {
-    "vscode": "^1.30.0"
-  },
-  "categories": [
-    "Programming Languages"
-  ],
-  "keywords": [
-    "scala",
-    "dotty",
-    "jvm",
-    "language-server"
-  ],
-  "main": "./out/src/extension",
-  "activationEvents": [
-    "onLanguage:scala",
-    "workspaceContains:.dotty-ide.json"
-  ],
-  "contributes": {
-    "languages": [
-      {
-        "id": "scala",
-        "extensions": [
-          ".scala",
-          ".sc"
-        ],
-        "aliases": [
-          "Scala"
-        ]
-      },
-      {
-        "id": "tasty",
-        "extensions": [
-          ".tasty"
-        ],
-        "aliases": [
-          "TASTy"
-        ]
-      }
-    ],
-    "configuration": {
-      "type": "object",
-      "title": "Dotty configuration",
-      "properties": {
-        "dotty.runWorksheetOnSave": {
-          "type": "boolean",
-          "default": true,
-          "description": "If true, saving a worksheet will also run it."
-        },
-        "dotty.trace.remoteWorkspaceDumpUrl": {
-          "type": "string",
-          "default": "",
-          "description": "Url to send local workspace dump to."
-        },
-        "dotty.trace.remoteTracingUrl": {
-          "type": "string",
-          "default": "",
-          "description": "Url to send local LSP communication to."
-        },
-        "dotty.trace.machineId": {
-          "type": [
-            "string",
-            "null"
-          ],
-          "default": null,
-          "description": "ID of your machine used when Dotty Language Server tracing is turned on."
-        },
-        "dotty.trace.maximumMessageSize": {
-          "type": "number",
-          "default": 512,
-          "description": "Maximum size for the messages sent to remote server. Larger ones will be split."
-        }
-      }
-    },
-    "commands": [
-      {
-        "command": "dotty.worksheet.run",
-        "title": "Run worksheet",
-        "category": "Scala"
-      },
-      {
-        "command": "dotty.worksheet.cancel",
-        "title": "Cancel running worksheet",
-        "category": "Scala"
-      }
-    ],
-    "configurationDefaults": {
-      "[scala]": {
-        "editor.tabSize": 2,
-        "editor.insertSpaces": true
-      }
-    },
-    "files.associations": {
-      "*.sc": "scala"
-    }
-  },
-  "scripts": {
-    "tsc": "./node_modules/.bin/tsc",
-    "vscode:prepublish": "npm install && ./node_modules/.bin/tsc -p ./",
-    "compile": "./node_modules/.bin/tsc -p ./",
-    "test": "node ./node_modules/vscode/bin/test",
-    "postinstall": "node ./node_modules/vscode/bin/install && curl -L -o out/coursier https://github.com/coursier/coursier/raw/v2.0.0-RC3-4/coursier"
-  },
-  "extensionDependencies": [
-    "scala-lang.scala"
-  ],
-  "dependencies": {
-    "archiver": "^3.1.1",
-    "compare-versions": "^3.5.1",
-    "promisify-child-process": "^2.1.2",
-    "request": "^2.88.0",
-    "vscode-jsonrpc": "4.0.0",
-    "vscode-languageclient": "^5.2.1",
-    "vscode-languageserver": "^5.2.1",
-    "ws": "^6.2.1"
-  },
-  "devDependencies": {
-    "@types/archiver": "^2.1.3",
-    "@types/compare-versions": "^3.0.0",
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^10.14.18",
-    "@types/request": "^2.48.3",
-    "@types/ws": "^6.0.3",
-    "typescript": "^3.6.3",
-    "vscode": "^1.1.37"
-  }
-}
