--- vscode-dotty/tsconfig.json
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es6",
-        "outDir": "out",
-        "lib": [
-            "es6"
-        ],
-        "sourceMap": true,
-        "rootDir": ".",
-        "strict": true
-    },
-    "exclude": [
-        "node_modules",
-        ".vscode-test"
-    ]
-}
